APP/requirements.txt
# ── Web Framework ──────────────────────────────────
flask>=3.0

# ── Elasticsearch ──────────────────────────────────
elasticsearch>=8.0,<9.0

# ── BERT / Sentence Transformers ───────────────────
sentence-transformers>=2.2
torch>=2.0

# ── FAISS (CPU version) ───────────────────────────
faiss-cpu>=1.7

# ── Utilities ─────────────────────────────────────
numpy>=1.24
